Don't need to pin managed objects
Internally a GCHandle is represented as an integer value. Getting an IntPtr from it is not taking the address of a managed object, it's merely a transformation of the GCHandle iteself. Therefore, the IntPtr is still an indirect reference and we don't have to worry about the managed object moving due to GC operations.